Separate GIF and image handling into dedicated services
